requirements.txt
fastapi
beanie
uvicorn
python-dotenv
pydantic-settings
passlib
bcrypt==4.0.1
pydantic[email]
python-multipart
python-jose